requirements.txt
pytest==8.2.1
setuptools==70.0.0
defusedxml==0.7.1
pyyaml==6.0.1
